pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alicia_d_sdk"
version = "6.1.0rc3"
description = "Python SDK for controlling the Alicia D robotic arm."
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Synria Robotics", email = "support@synriarobotics.ai"}
]
keywords = ["robotics", "robotic-arm", "alicia-d", "sdk", "serial-communication"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware :: Hardware Drivers",
    "Topic :: Scientific/Engineering", 
]

dependencies = [
    "pyserial>=3.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.3.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "websockets==13.1",
    "pythoncrc",
    "pycrc",
    "synriard>=1.0.3rc1",  
    "synria-robocore>=2.0.0alpha1",
]

[project.urls]
Homepage = "https://github.com/Synria-Robotics/Alicia-D-SDK"

[tool.setuptools]
# Automatically find all packages and subpackages, excluding examples
packages = {find = {exclude = ["examples*"]}}
# Set to true so config files inside the package are included
include-package-data = true